requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-cors-headers
django-extensions
django-filter
djangorestframework
djangorestframework_simplejwt
djoser
drf-yasg
idna
inflection
oauthlib
packaging
pillow
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
PyYAML
requests
requests-oauthlib
rest-framework-simplejwt
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
uritemplate
urllib3
